Big fixes, amelioration and a lot of debug and documentation reading...

+++ src/errors/raiseError.ts
@@ -1,3 +1,4 @@
+import { ICM } from "../interpreter/interpreter.js"
 import CompilerError from "./_error.js"
 import UnknowError from "./compiler/UnknowError.js"
 
@@ -10,7 +11,7 @@ export default class RaiseFlyLangCompilerError {
     raise() {
         console.error(this.error.toString())
         
-        process.kill(process.pid) // This kill the current process (without any additionnal message except the error's one)
+        if(!ICM()) process.kill(process.pid) // This kill the current process (without any additionnal message except the error's one)
         return null // For typing (when the code above is executed, the process is killed and stop)
     }
 }
